refactor: move citis liist into its own component
